server/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
google-api-core==1.19.1
google-auth==1.16.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.9
numpy==1.18.5
pandas==1.0.4
pg8000==1.15.2
postgres==3.0.0
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
scramp==1.1.1
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
xlrd==1.2.0

